feat: Warn about no route and move files
