+++ packages/prediction-app/src/features/common-features/WarnRouteNotExists/WarnRouteNotExists.module.css
@@ -1,0 +1,9 @@
+.warningMargin {
+    /*margin-block-start: 24px;*/
+    margin: 0 auto;
+}
+
+.warningMarginInner {
+    margin-top: 10px;
+    margin-bottom: 10px;
+}
