+++ packages/prediction-app/src/features/new-prediction/NewPredictionDrawer.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import Drawer from 'react-modern-drawer'
 import 'react-modern-drawer/dist/index.css'
-import PredictionPage from '../../components/prediction/PredictionPage'
 import styles from './NewPredictionDrawer.module.css'
 import NewPredictionForm from './NewPredictionForm/NewPredictionForm'
