--- packages/prediction-app/src/features/settings/WarnRouteNotExists.module.css
@@ -1,3 +1,0 @@
-.warningMargin {
-    margin-block-start: 24px;
-}
